Restore RG compat - also clean up some code - update manifest

+++ scripts/_promptListeners.mjs
@@ -1,50 +1,44 @@
 import {MODULE} from "./settings.mjs";
 
-export function _clickPrompt(_, html) {
-  html[0].addEventListener("click", (event) => {
-    return clickConcentrationPrompt(event);
-  });
+/**
+ * Hook function to append listener onto rendered chat messages.
+ * @param {ChatMessage} message     The message rendered.
+ * @param {html} html               The element of the message.
+ */
+export function _clickPrompt(message, html) {
+  html[0].querySelector(".concentrationnotifier [data-prompt='saving-throw']")?.addEventListener("click", _onClickSavingThrow);
+  html[0].querySelector(".concentrationnotifier [data-prompt='end-concentration']")?.addEventListener("click", _onClickEndConcentration);
+  html[0].querySelector(".concentrationnotifier [data-prompt='remove-templates']")?.addEventListener("click", _onClickRemoveTemplates);
+  html[0].querySelector(".concentrationnotifier [data-prompt='render-sheet']")?.addEventListener("click", _onClickRenderSheet);
 }
 
-async function clickConcentrationPrompt(event) {
-  // get the target of the mouse click.
-  let button = event.target?.closest(".concentrationnotifier .buttons > button");
-  if (!button) button = event.target?.closest("[name='render-item-sheet']");
-  if (!button) return;
-
-  if (button.name === "saving-throw") {
-    const caster = fromUuidSync(button.dataset.actorUuid);
-    const actor = caster.actor ?? caster;
-    return actor.rollConcentrationSave(button.dataset.saveType, {
-      targetValue: button.dataset.dc
-    });
-  }
-  else if (button.name === "delete-concentration") {
-    const effect = fromUuidSync(button.dataset.effectUuid);
-    return deleteDialog(effect, event);
-  }
-  else if (button.name === "remove-templates") {
-    const templateIds = canvas?.scene.templates.filter(t => {
-      return t.isOwner && t.flags.dnd5e?.origin === button.dataset.origin;
-    }).map(t => t.id);
-    return canvas?.scene.deleteEmbeddedDocuments("MeasuredTemplate", templateIds);
-  }
-  else if (button.name === "render-item-sheet") {
-    const item = fromUuidSync(button.dataset.itemUuid);
-    return item.sheet.render(true);
-  }
+/**
+ * Perform a saving throw to maintain concentration on the spell.
+ * @param {PointerEvent} event      The initiating click event.
+ * @returns {Promise<D20Roll>}      The rolled save.
+ */
+function _onClickSavingThrow(event) {
+  const data = event.currentTarget.dataset;
+  const caster = fromUuidSync(data.actorUuid);
+  const actor = caster.actor ?? caster;
+  return actor.rollConcentrationSave(data.saveType, {targetValue: data.dc});
 }
 
-async function deleteDialog(effect, event) {
-  if (!effect) return;
-  if (event.shiftKey) {
-    return effect.delete();
-  }
+/**
+ * Prompt or immediately end the concentration.
+ * @param {PointerEvent} event                  The initiating click event.
+ * @returns {Promise<ActiveEffect>|Dialog}      The deleted effect, or the rendered prompt.
+ */
+function _onClickEndConcentration(event) {
+  const effect = fromUuidSync(event.currentTarget.dataset.effectUuid);
+  if (event.shiftKey) return effect.delete();
+
   const name = effect.flags[MODULE].data.itemData.name;
-  new Dialog({
-    title: game.i18n.format("CN.ConfirmEndConcentrationTitle", {name}),
-    content: game.i18n.format("CN.ConfirmEndConcentrationText", {name}),
-    buttons: {
+  const title = game.i18n.format("CN.ConfirmEndConcentrationTitle", {name});
+  const content = game.i18n.format("CN.ConfirmEndConcentrationText", {name});
+  const id = `${MODULE}-deletePrompt-${effect.uuid.replaceAll(".", "-")}`;
+  return new Dialog({
+    title, content, buttons: {
       yes: {
         icon: "<i class='fa-solid fa-check'></i>",
         label: game.i18n.localize("Yes"),
@@ -57,5 +51,26 @@ async function deleteDialog(effect, event) {
         label: game.i18n.localize("No")
       }
     }
-  }).render(true);
+  }, {id}).render(true);
+}
+
+/**
+ * Delete all related templates on the scene that you have permission to delete.
+ * @param {PointerEvent} event                The initiating click event.
+ * @returns {MeasuredTemplateDocument[]}      The deleted templates.
+ */
+function _onClickRemoveTemplates(event) {
+  const ids = canvas?.scene.templates.filter(t => {
+    return t.isOwner && (t.flags.dnd5e?.origin === event.currentTarget.dataset.origin);
+  }).map(t => t.id);
+  return canvas?.scene.deleteEmbeddedDocuments("MeasuredTemplate", ids);
+}
+
+/**
+ * Render the item being concentrated on.
+ * @param {PointerEvent} event      The initiating click event.
+ * @returns {ItemSheet5e}           The rendered item sheet.
+ */
+function _onClickRenderSheet(event) {
+  return fromUuidSync(event.currentTarget.dataset.itemUuid).sheet.render(true);
 }
